Issue #4973 relocated bundle routes to their own file
