+++ backend/src/v5/routes/routesManager.js
@@ -17,6 +17,7 @@
 
 const RoutesManager = {};
 const AadRoutes = require('./sso/aad');
+const BundleRoutes = require('./teamspaces/projects/models/containers/bundles');
 const CalibrationRoutes = require('./teamspaces/projects/models/drawings/calibrations');
 const CreateGeneralRevisionRoutes = require('./teamspaces/projects/models/common/revisions');
 const CreateGroupRoutes = require('./teamspaces/projects/models/common/groups');
@@ -57,6 +58,7 @@ RoutesManager.init = (app) => {
 	app.use('/v5/teamspaces/:teamspace/projects/:project/containers/:model/views', CreateViewRoutes());
 	app.use('/v5/teamspaces/:teamspace/projects/:project/containers/:model/revisions', CreateGeneralRevisionRoutes(modelTypes.CONTAINER));
 	app.use('/v5/teamspaces/:teamspace/projects/:project/containers/:container/metadata', MetadataRoutes);
+	app.use('/v5/teamspaces/:teamspace/projects/:project/containers/:container/bundle', BundleRoutes);
 
 	// Federations
 	app.use('/v5/teamspaces/:teamspace/projects/:project/federations', CreateModelGeneralRoutes(modelTypes.FEDERATION));
